python/requirements.txt
flask_cors
flask
qqmusic-api-python
httpx
asyncio